+++ db/access.php
@@ -348,4 +348,13 @@ $capabilities = [
             'manager' => CAP_ALLOW,
         ],
     ],
+    // Ability to override the course header settings in a course.
+    'theme/boost_union:overridecourseheaderincourse' => [
+        'captype' => 'write',
+        'contextlevel' => CONTEXT_COURSE,
+        'archetypes' => [
+            'editingteacher' => CAP_ALLOW,
+            'manager' => CAP_ALLOW,
+        ],
+    ],
 ];