+++ db/upgrade.php
@@ -691,6 +691,62 @@ function xmldb_theme_boost_union_upgrade($oldversion) {
         upgrade_plugin_savepoint(true, 2025041419, 'theme', 'boost_union');
     }
 
+    if ($oldversion < 2025041428) {
+        // Define table theme_boost_union_course to be created.
+        $table = new xmldb_table('theme_boost_union_course');
+
+        // Adding fields to table.
+        $table->add_field('id', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, XMLDB_SEQUENCE, null);
+        $table->add_field('courseid', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, null);
+        $table->add_field('name', XMLDB_TYPE_CHAR, '100', null, XMLDB_NOTNULL, null, null);
+        $table->add_field('value', XMLDB_TYPE_TEXT, null, null, null, null, null);
+
+        // Adding keys to table.
+        $table->add_key('primary', XMLDB_KEY_PRIMARY, ['id']);
+        $table->add_key('courseid', XMLDB_KEY_FOREIGN, ['courseid'], 'course', ['id']);
+
+        // Adding indexes for performance.
+        $table->add_index('courseidname', XMLDB_INDEX_UNIQUE, ['courseid', 'name']);
+
+        // Create the table if it doesn't exist.
+        if (!$dbman->table_exists($table)) {
+            $dbman->create_table($table);
+        }
+
+        // Additionally, rename multiple course header settings.
+        $oldvalue = get_config('theme_boost_union', 'courseheaderimageenabled');
+        if ($oldvalue !== false) {
+            set_config('courseheaderenabled', $oldvalue, 'theme_boost_union');
+            unset_config('courseheaderimageenabled', 'theme_boost_union');
+        }
+        $oldvalue = get_config('theme_boost_union', 'courseheaderimagelayout');
+        if ($oldvalue !== false) {
+            set_config('courseheaderlayout', $oldvalue, 'theme_boost_union');
+            unset_config('courseheaderimagelayout', 'theme_boost_union');
+        }
+        $oldvalue = get_config('theme_boost_union', 'courseheaderimageheight');
+        if ($oldvalue !== false) {
+            set_config('courseheaderheight', $oldvalue, 'theme_boost_union');
+            unset_config('courseheaderimageheight', 'theme_boost_union');
+        }
+        $oldvalue = get_config('theme_boost_union', 'courseheaderimagefallback');
+        if ($oldvalue !== false) {
+            set_config('courseheaderimageglobal', $oldvalue, 'theme_boost_union');
+            unset_config('courseheaderimagefallback', 'theme_boost_union');
+        }
+
+        // Move any existing file in the courseheaderimagefallback filearea to the new courseheaderimageglobal filearea.
+        // We do this just with SQL in the mdl_files table and do not go through the file API.
+        $query = "UPDATE {files}
+                  SET filearea = 'courseheaderimageglobal'
+                  WHERE component = 'theme_boost_union'
+                  AND filearea = 'courseheaderimagefallback'";
+        $DB->execute($query);
+
+        // Boost Union savepoint reached.
+        upgrade_plugin_savepoint(true, 2025041428, 'theme', 'boost_union');
+    }
+
     // Load the builtin SCSS snippets into the database.
     // This is done with every plugin update, regardless of the plugin version.
     snippets::add_builtin_snippets();