+++ lang/en/theme_boost_union.php
@@ -41,6 +41,11 @@ $string['auto'] = 'Automatically';
 $string['nochange'] = 'No change';
 $string['forguestsonly'] = 'Only for guests and non-logged-in users';
 
+// Course overrides: General strings.
+$string['courseoverride'] = 'Add to course settings as well';
+$string['courseoverride_desc'] = 'The Boost Union settings in this section come with the possibility to override the global setting (which is set here) in each course. As soon as you tick the \'Add to course settings as well\' checkbox for a setting, the setting will be added to the course settings form where teachers, managers and admins (or any other role which you add the theme/boost_union:overridecourseheaderincourse capability to) can override the global setting for each course individually. If you do not tick the checkbox, the setting will not be added to the course settings form and the global setting will always apply.';
+$string['nocourseoverride'] = 'This setting cannot be overridden in the course settings.';
+
 // Settings: General strings.
 $string['dontchange'] = 'Do not change anything';
 $string['tertiarysettings'] = 'List of Boost Union settings pages';
@@ -429,21 +434,46 @@ $string['upcomingeventstintenabled_desc'] = 'With this setting, you can tint the
 // Settings: Course tab.
 $string['coursetab'] = 'Course';
 // ... Section: Course header.
-$string['courseheaderheading'] = 'Course Header';
+$string['courseheaderheading'] = 'Course header';
 // ... ... Setting: Course header.
-$string['courseheaderimageenabled'] = 'Display the course image in the course header';
-$string['courseheaderimageenabled_desc'] = 'When enabled, the course image (which can be uploaded in a course\'s course settings) is displayed in the header of a course. The course images are shown there in addition to the \'My courses\' page where they are always shown.';
-$string['courseheaderimagefallback'] = 'Fallback course header image';
-$string['courseheaderimagefallback_desc'] = 'If you upload an image in this setting, it is used as fallback image and is displayed in the course header if no course image is uploaded in a particular course\'s course settings. If you do not upload an image here, a course header image is only shown in a particular course if a course image is uploaded in this particular course\'s course settings.';
-$string['courseheaderimageheight'] = 'Course header image height';
-$string['courseheaderimageheight_desc'] = 'With this setting, you control the height of the presented course header image.';
-$string['courseheaderimagelayout'] = 'Course header image layout';
-$string['courseheaderimagelayout_desc'] = 'With this setting, you control the layout of the course header image and the course title.';
-$string['courseheaderimagelayoutstackeddark'] = 'Course title stacked on course image (white font color for dark background images)';
-$string['courseheaderimagelayoutstackedlight'] = 'Course title stacked on course image (black font color for light background images)';
-$string['courseheaderimagelayoutheadingabove'] = 'Course title above of course image';
+$string['courseheaderenabled'] = 'Enable enhanced course header';
+$string['courseheaderenabled_desc'] = 'When enabled, the course header (which is just the course title in Moodle core) is enhanced by additional elements like the course image (which can be uploaded in a course\'s course settings) and other course metadata, depending how you configure the course header in detail.';
+$string['courseheaderenabled_help'] = 'When enabled, the course header is enhanced by additional elements like the course image and other course metadata, depending how you configure the course header in detail.';
+$string['courseheaderheight'] = 'Course header height';
+$string['courseheaderheight_desc'] = 'With this setting, you control the height of the presented course header.';
+$string['courseheaderheight_help'] = 'The height of the presented course header.';
+$string['courseheaderlayout'] = 'Course header layout';
+$string['courseheaderlayout_desc'] = 'With this setting, you control the layout of the course header and the course title.';
+$string['courseheaderlayout_help'] = 'The layout of the course header and the course title.';
+$string['courseheaderlayoutstackeddark'] = 'Course title stacked on course header image (white font color for dark background images)';
+$string['courseheaderlayoutstackedlight'] = 'Course title stacked on course header image (black font color for light background images)';
+$string['courseheaderlayoutheadingabove'] = 'Course title above of course header image';
+$string['courseheaderimageglobal'] = 'Global course header image';
+$string['courseheaderimageglobal_desc'] = 'The image which you upload here is used as global course header image in the cascade which you configure in the \'Course header image source\' setting.';
 $string['courseheaderimageposition'] = 'Course header image position';
 $string['courseheaderimageposition_desc'] = 'With this setting, you control the positioning of the course header image within the course header image container. The first value is the horizontal position, the second value is the vertical position.';
+$string['courseheaderimageposition_help'] = 'The positioning of the course header image within the course header image container. The first value is the horizontal position, the second value is the vertical position.';
+$string['courseheaderimagesource'] = 'Course header image source';
+$string['courseheaderimagesource_desc'] = 'Several course header layouts contain the presentation of a course header image. With this setting, you control the source of that course header image.';
+$string['courseheaderimagesource_explanation'] = 'Explanation of terms: <br /><ul><li><em>Course image</em>: The course image which can be uploaded in a particular course\'s course settings based on Moodle core functionality.</li><li><em>Global course header image</em>: The image which can be uploaded in the \'Global course header image\' setting below.</li><li><em>Dedicated course header image</em>: A dedicated course header image which can be uploaded in a particular course\'s course settings. As soon as you configure this setting to an option which contains a dedicated course header image, an additional file upload setting appears in each course\'s settings page. Additionally, to improve the structure of the course settings, it is combined with the Moodle Core \'Course image\' setting under a new \'Course images\' course settings section.</li></ul>';
+$string['courseheaderimagesource_courseplusglobal'] = 'Course image with a fallback to the global course header image';
+$string['courseheaderimagesource_coursenoglobal'] = 'Course image without fallback to the global course header image';
+$string['courseheaderimagesource_dedicatedplusfallback'] = 'Dedicated course header image with a fallback to the global course header image';
+$string['courseheaderimagesource_dedicatednofallback'] = 'Dedicated course header image without fallback to the global course header image';
+$string['courseheaderimagesource_global'] = 'Global course header image for all courses';
+$string['courseimagesheading'] = 'Course images';
+$string['courseheaderimage'] = 'Course header image';
+$string['courseheaderimageplusfallback'] = 'Course header image'; // This string is just there to make the help_icon class happy but is not used anywhere.
+$string['courseheaderimagenofallback'] = 'Course header image'; // This string is just there to make the help_icon class happy but is not used anywhere.
+$string['courseheaderimageplusfallback_help'] = 'The course header image is displayed in the course header at the top of the course page. If you do not upload an image here, a global course header image will be presented.';
+$string['courseheaderimagenofallback_help'] = 'The course header image is displayed in the course header at the top of the course page. If you do not upload an image here, the course header will – depending on the configured course header layout – just show the course title or at least be displayed with a reduced layout.';
+$string['courseheaderlayoutexclusionlist'] = 'Course header layouts exclusion list';
+$string['courseheaderlayoutexclusionlist_desc'] = 'With this setting, you can exclude particular course header layouts from being available for course-specific overrides. The selected layouts will not appear in the course settings for teachers to choose from, but the administrator can still use them globally. This allows you to restrict which layouts are available to course editors while maintaining administrative control over the global setting. Please note that the course header layout which is selected by the admin above will never be excluded from the course settings, even if you tick it here.';
+$string['courseheaderformatexclusionlist'] = 'Course format exclusion list';
+$string['courseheaderformatexclusionlist_desc'] = 'With this setting, you can exclude particular course formats from the course header feature. For courses using the selected course formats, the course header will not be modified regardless of the other course header settings. This allows you to exclude course formats which do not work well with the course header feature or which have their own course header implementation.';
+$string['courseheaderimporttransfer'] = 'Transfer course-specific header settings during course import';
+$string['courseheaderimporttransfer_desc'] = 'With this setting, you can control whether course header settings and the course header image which have been set within a particular course\'s settings are transferred to another course within a course import. If set to "Always", all course header settings will be copied from the source course to the destination course, regardless if the teacher imports a whole course or just a fraction of a course. If set to "Never", course header settings will not be transferred. Allowing the teacher to decide this question during the import process is not possible due to technical limitations of Moodle core.';
+$string['courseheaderrestoreoption'] = 'Include course header settings';
 // ... Section: Course index.
 $string['courseindexheading'] = 'Course Index';
 // ... ... Setting: Course index.
@@ -1636,6 +1666,7 @@ $string['boost_union:viewregionoffcanvasright'] = 'To be able to see the Off-can
 $string['boost_union:editregionoffcanvasright'] = 'To be able to edit the Off-canvas (right) block region';
 $string['boost_union:viewregionoffcanvascenter'] = 'To be able to see the Off-canvas (center) block region';
 $string['boost_union:editregionoffcanvascenter'] = 'To be able to edit the Off-canvas (center) block region';
+$string['boost_union:overridecourseheaderincourse'] = 'To be able to override the course header settings in a course';
 
 // Caches.
 $string['cachedef_flavours'] = 'Flavours which apply to a given page\'s category ID for the current user';
@@ -1644,6 +1675,7 @@ $string['cachedef_smartmenu_items'] = 'Smart menu items';
 $string['cachedef_touchiconsios'] = 'Touch icon files for iOS';
 $string['cachedef_hookoverrides'] = 'Hook overrides';
 $string['cachedef_fontawesomeicons'] = 'FontAwesome icon map';
+$string['cachedef_courseoverrides'] = 'Course-specific setting overrides';
 
 // Scheduled tasks.
 $string['task_purgecache'] = 'Purge theme cache';